Initial attempt to break media carousel out separately from viewer.
