+++ src/scss/media-carousel.scss
@@ -6,7 +6,7 @@
   flex: 0 0 100%;
 }
 
-frigate-card-ha-hls-player {
+* {
   height: 100%;
   width: 100%;
   transform-style: preserve-3d;
