+++ src/scss/viewer-core.scss
@@ -7,7 +7,7 @@
   width: 100%;
 }
 
-frigate-card-media-carousel {
+frigate-card-viewer-carousel {
   flex: 1;
   min-height: 0;
 }
